logout fixed
